Düzeltme.

+++ CareGardenApiV1/Repository/Concrete/BusinessRepository.cs
@@ -169,7 +169,7 @@ namespace CareGardenApiV1.Repository.Concrete
             var filteredBusinesses = businesses
                 .WhereIf(businessExploreModel.workingGenderType != WorkingGenderType.All, x => x.workingGenderType.Equals((int)businessExploreModel.workingGenderType))
                 .WhereIf(businessExploreModel.serviceId.HasValue, x => x.serviceIds != null && x.serviceIds.Contains(businessExploreModel.serviceId))
-                .WhereIf(!businessExploreModel.offers.IsNullOrEmpty(), x => x.discounts != null && x.discounts.Exists(x => x.isActive && businessExploreModel.offers.Contains(x.rate))
+                .WhereIf(!businessExploreModel.offers.IsNullOrEmpty(), x => x.discounts != null && x.discounts.Exists(x => x.isActive && businessExploreModel.offers.Contains((int)x.rate)))
                 .WhereIf(businessExploreModel.availableDate.HasValue, x => HelperMethods.IsAvailableAppointmentDay(x.appointments, x.workingInfo, x.officialDayAvailable, businessExploreModel.availableDate.Value))
                 .WhereIf(!businessExploreModel.city.IsNullOrEmpty(), x => x.city.Equals(businessExploreModel.city))
                 .Select(x =>
